Refactor file names to follow naming conventions
